* commit 'dbfb1ea964dbb853abb59b69636d918bc6b234ef':
  ECKIT-377 fix compiler warnings
  Remove using statements,
  Removed variable rather than comment
  Fix 2 more warnings
  Fix 100ds of warnings in eckit_sql
  Fix numerous compilation warnings

+++ src/eckit/log/Number.cc
@@ -59,7 +59,7 @@ void Number::printBinary(std::ostream& ss, W n)
         }
     }
     else
-    for (int i = 0; i < sizeof(W); ++i)
+    for (size_t i = 0; i < sizeof(W); ++i)
     {
         unsigned char c = s[i];
         for(unsigned char mask = 1 << 7; mask; mask >>= 1)
